fix:sdk removeWarnings and move StableCoinNotAssociated error
