+++ sdk/src/app/usecase/command/stablecoin/error/StableCoinNotAssociated.ts
@@ -1,4 +1,4 @@
-import BaseError, { ErrorCode } from '../../../../core/error/BaseError.js';
+import BaseError, { ErrorCode } from '../../../../../core/error/BaseError.js';
 
 export class StableCoinNotAssociated extends BaseError {
 	constructor(targetId: unknown, tokenId: unknown) {
